model files separated
